chore(AutoDocs): refactor - split props and methods table
also, relocate depreacted props logic from api section to AutoDocs
component
