+++ packages/wix-storybook-utils/src/AutoDocs/styles.scss
@@ -1,7 +1,23 @@
 @import '../ui/typography.scss';
+@import '../ui/colors.scss';
 
 .tableTitle {
   @include font(medium);
-  margin: 20px 0;
+  margin-bottom: 20px;
   font-size: 18px;
+}
+
+.propsTable {
+  margin-bottom: 60px;
+}
+
+.deprecatedTitle {
+  @include font(medium);
+  margin-bottom: 20px;
+  font-size: 18px;
+}
+
+.deprecatedDescription {
+  color: $R00;
+  margin-bottom: 20px;
 }
