--- packages/wix-storybook-utils/src/Sections/views/api/styles.scss
@@ -1,13 +1,0 @@
-@import '../../../ui/typography.scss';
-@import '../../../ui/colors';
-
-.deprecatedTitle {
-  margin: 30px 0;
-  @include font(bold);
-}
-
-.deprecatedDescription {
-  margin: 10px 0;
-  @include font(light);
-  color: $R10;
-}
